feature: ephemeral public key extension
